+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.10.98 fix: fix get_cffex_rank_table interface
+
+    1. 修复 get_cffex_rank_table 接口
+
 1.10.97 fix: fix stock_financial_analysis_indicator interface
     
     1. 修复 stock_financial_analysis_indicator 接口
@@ -2820,6 +2824,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.98 fix: fix get_cffex_rank_table interface
 
 1.10.97 fix: fix stock_financial_analysis_indicator interface